feat(tests): add test-shell binary and update test helpers for improved execution
