+++ main_test.go
@@ -1,10 +1,8 @@
 package main
 
 import (
-	"context"
 	"fmt"
 	"os"
-	"os/exec"
 	"path/filepath"
 	"strings"
 	"testing"
@@ -22,31 +20,23 @@ var (
 )
 
 func TestMain(m *testing.M) {
-	// using an anonymous function to handle defer statements before os.Exit()
 	exitCode := func() int {
-		ctx, cancel := context.WithTimeout(context.Background(), constants.DefaultTestTimeout)
-		defer cancel()
-
-		tempDir, err := os.MkdirTemp("", "resticprofile-main")
-		if err != nil {
-			fmt.Fprintf(os.Stderr, "cannot create temp dir: %v\n", err)
-			return 1
-		}
-		fmt.Printf("using temporary dir: %q\n", tempDir)
-		defer os.RemoveAll(tempDir)
-
-		mockBinary = filepath.Join(tempDir, platform.Executable("shellmock"))
-
-		cmdMock := exec.CommandContext(ctx, "go", "build", "-buildvcs=false", "-o", mockBinary, "./shell/mock")
-		if output, err := cmdMock.CombinedOutput(); err != nil {
-			fmt.Fprintf(os.Stderr, "Error building shell/mock binary: %s\nCommand output: %s\n", err, string(output))
-			return 1
-		}
-
+		var err error
 		helpersPath := os.Getenv("TEST_HELPERS")
 		if helpersPath == "" {
 			helpersPath = "./build"
 		}
+		mockBinary = filepath.Join(helpersPath, platform.Executable("test-shell"))
+		mockBinary, err = filepath.Abs(mockBinary)
+		if err != nil {
+			fmt.Fprintf(os.Stderr, "Failed to get absolute path of test-shell binary: %v\n", err)
+			return 1
+		}
+		if _, err := os.Stat(mockBinary); err != nil {
+			fmt.Fprintf(os.Stderr, "test-shell binary is not available at expected path: %s\n", mockBinary)
+			return 1
+		}
+
 		echoBinary = filepath.Join(helpersPath, platform.Executable("test-echo"))
 		echoBinary, err = filepath.Abs(echoBinary)
 		if err != nil {
