+++ testhelpers/args/priority_unix.go
@@ -1,4 +1,4 @@
-//go:build linux
+//go:build !windows && !linux
 
 package main
 
@@ -7,23 +7,21 @@ import (
 	"os"
 	"sync"
 
-	"github.com/creativeprojects/resticprofile/priority"
 	"golang.org/x/sys/unix"
 )
 
 const selfPID = 0
 
-func main() {
+// This is only displaying the priority of the current process (for testing)
+func runPriority() int {
 	// run it in a go routine in case it would make a difference
 	wg := sync.WaitGroup{}
-	wg.Add(1)
-	go func() {
-		defer wg.Done()
+	wg.Go(func() {
 		displayProcessAndGroup()
 		displayPriority()
-		displayIOPriority()
-	}()
+	})
 	wg.Wait()
+	return 0
 }
 
 func displayProcessAndGroup() {
@@ -50,7 +48,7 @@ func getProcessNice() (int, error) {
 	if err != nil {
 		return 0, err
 	}
-	return 20 - pri, nil
+	return pri, nil
 }
 
 // GetProcessNice returns the nice value of the current process group
@@ -59,14 +57,5 @@ func getGroupNice() (int, error) {
 	if err != nil {
 		return 0, err
 	}
-	return 20 - pri, nil
-}
-
-func displayIOPriority() {
-	class, value, err := priority.GetIONice()
-	if err != nil {
-		fmt.Println(err)
-		return
-	}
-	fmt.Printf("IOPriority: class = %d, value = %d\n", class, value)
+	return pri, nil
 }
