+++ testhelpers/generate_unix.go
@@ -5,3 +5,4 @@ package testhelpers
 //go:generate go build -o ../build/test-args -buildvcs=false ./args
 //go:generate go build -o ../build/test-echo -buildvcs=false ./echo
 //go:generate go build -o ../build/test-crontab -buildvcs=false ./crontab
+//go:generate go build -o ../build/test-shell -buildvcs=false ./shell
